More Robust Version
In this new version, I do many experiment to adjust the funtion
parameters, and make my output more stable and smooth. Finally it can
creat a avi. video(without sound) as output and crudely pull the nose to
the middle of the video(but the nose not in the middle all the way!!).
And this repository will be continually updated...